+++ mocks/auth/main.js
@@ -22,7 +22,10 @@ module.exports = function(http) {
 
             this.respond(200, (userCache.user = extend(user, {
                 email: request.body.email,
-                id: userId
+                id: userId,
+                entitlements: {
+                    adminCampaigns: true
+                }
             })));
         } else {
             this.respond(401, 'Invalid email or password');